# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lilypad API",
        "description": "API for lilypad",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/payment-terms": {
            "get": {
                "tags": [
                    "Payment Terms"
                ],
                "summary": "Get payment terms by name or all payment terms",
                "operationId": "68c882f8a269b2e2e19984a6d9532485",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of the payment terms to retrieve",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the payment terms to retrieve",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment Terms retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Payment Terms retrieved successfully!"
                                        },
                                        "data": {
                                            "description": "Payment Terms object",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment Terms not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Payment Terms not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Payment Terms"
                ],
                "summary": "Create new payment terms",
                "description": "Creates new payment terms with provided details such as name, type, net days, discount, etc.",
                "operationId": "61c1e40f4832ba0a64d84880ecc46438",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "termsName": {
                                        "description": "Name of the payment terms",
                                        "type": "string",
                                        "example": "Net 30"
                                    },
                                    "termsType": {
                                        "description": "Type of payment terms",
                                        "type": "string",
                                        "example": "CCD"
                                    },
                                    "netDays": {
                                        "description": "Net days for the payment",
                                        "type": "integer",
                                        "example": 30
                                    },
                                    "discount": {
                                        "description": "Discount percentage",
                                        "type": "number",
                                        "format": "float",
                                        "example": 5
                                    },
                                    "discountDays": {
                                        "description": "Days to apply the discount",
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "dueDate": {
                                        "description": "Due date for the payment",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-09-01"
                                    },
                                    "nextMonth": {
                                        "description": "Next month's date",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-09-01"
                                    },
                                    "discountDate": {
                                        "description": "Discount date",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-08-15"
                                    },
                                    "default": {
                                        "description": "Flag for default terms",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "active": {
                                        "description": "Flag for active status",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Payment Terms Created Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Payment Terms Created Successfully!"
                                        },
                                        "data": {
                                            "description": "Created payment terms object",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "description": "Validation errors",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}